rename: 폴더 store -> redux 로 이름 변경
