+++ client/src/redux/store.ts
@@ -1,9 +1,11 @@
 import { configureStore } from '@reduxjs/toolkit';
 import authSlice from './authSlice';
+import  questionReducer  from './questionReducer';
 
 export const store = configureStore({
   reducer: {
     auth: authSlice,
+    questionReducer
   },
 });
